refactor: template 하위 디렉토리 설정
