+++ components/templates/home-template/HomeTemplate.stories.tsx
@@ -1,0 +1,11 @@
+import HomeTemplate from './HomeTemplate';
+
+export default {
+    title: 'Design Systems/Templates/Home-Template',
+    component: HomeTemplate,
+    tags: ['autodocs'],
+};
+
+export const homeTemplate = () => {
+    return <HomeTemplate></HomeTemplate>;
+};
